+++ nodes/Discord/bot/index.ts
@@ -62,9 +62,46 @@ export default function () {
 
   // listen to users changing their status events
   client.on('presenceUpdate', (oldPresence, newPresence) => {
-    console.log('presenceUpdate', oldPresence, newPresence);
-    // userId, status (online, offline, idle, dnd)
-    // get roles ?
+    const member = newPresence.member;
+    try {
+      if (!member || member.user.bot || member.user.system) return;
+      const userRoles = member.roles.cache.map((role) => role.id);
+      Object.keys(state.channels).forEach((key) => {
+        const channel = state.channels[key];
+        channel.forEach(async (trigger) => {
+          if (trigger.roleIds.length) {
+            const hasRole = trigger.roleIds.some((role) => userRoles?.includes(role));
+            if (!hasRole) return;
+          }
+          if (
+            trigger.type === 'userPresenceUpdate' &&
+            (trigger.presence === newPresence.status || trigger.presence === 'any')
+          ) {
+            addLog(`triggerWorkflow ${trigger.webhookId}`, client);
+            const placeholderMatchingId = trigger.placeholder ? uid() : '';
+            const isEnabled = await triggerWorkflow(
+              trigger.webhookId,
+              null,
+              placeholderMatchingId,
+              state.baseUrl,
+              member.user,
+              key,
+              newPresence.status,
+            ).catch((e) => e);
+            if (isEnabled && trigger.placeholder) {
+              const channel = client.channels.cache.get(key);
+              const placeholder = await (channel as TextChannel)
+                .send(trigger.placeholder)
+                .catch((e: any) => addLog(`${e}`, client));
+              if (placeholder)
+                placeholderLoading(placeholder, placeholderMatchingId, trigger.placeholder);
+            }
+          }
+        });
+      });
+    } catch (e) {
+      addLog(`${e}`, client);
+    }
   });
 
   client.on('guildMemberUpdate', (oldMember, newMember) => {
@@ -83,7 +120,6 @@ export default function () {
         const channel = state.channels[key];
         channel.forEach(async (trigger) => {
           if (trigger.type === 'userJoins') {
-            // key
             addLog(`triggerWorkflow ${trigger.webhookId}`, client);
             const placeholderMatchingId = trigger.placeholder ? uid() : '';
             const isEnabled = await triggerWorkflow(
@@ -123,7 +159,6 @@ export default function () {
             if (!hasRole) return;
           }
           if (trigger.type === 'userLeaves') {
-            // key
             addLog(`triggerWorkflow ${trigger.webhookId}`, client);
             const placeholderMatchingId = trigger.placeholder ? uid() : '';
             const isEnabled = await triggerWorkflow(